% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfBands.R
\name{confbands}
\alias{confbands}
\title{Confidence Bands}
\usage{
confbands(lmObject, xh = NULL, alpha = 0.05)
}
\arguments{
\item{alpha}{level}

\item{lm}{object}
}
\value{
graph of confidence bands or interval
}
\description{
This function will use the lm function to create and add confidence bands to a previous graph.
With a given parameter it can also out put the interval of those bands at that parameter
}
